add independent testserver
