+++ .cspell.json
@@ -69,6 +69,7 @@
         "supersedable",
         "Tantek",
         "testtest",
+        "testserver",
         "tocline",
         "tocxref",
         "ttml",
