+++ test/rules.js
@@ -3,14 +3,9 @@
 import { expect as chai } from 'chai';
 // eslint-disable-next-line node/no-unpublished-import
 import expect from 'expect.js';
-// External packages:
-import express from 'express';
-import exphbs from 'express-handlebars';
-import { lstatSync, readdirSync } from 'fs';
+
 // eslint-disable-next-line node/no-unpublished-import
 import nock from 'nock';
-import pth, { dirname } from 'path';
-import { fileURLToPath } from 'url';
 import { Sink } from '../lib/sink.js';
 import { allProfiles } from '../lib/util.js';
 // Internal packages:
@@ -19,9 +14,8 @@ import { Specberus } from '../lib/validator.js';
 // Shouldn't cause any error.
 import { goodDocuments } from './data/goodDocuments.js';
 import { samples } from './samples.js';
-
-// eslint-disable-next-line no-underscore-dangle
-const __dirname = dirname(fileURLToPath(import.meta.url));
+import { app } from './lib/testserver.js';
+import { buildBadTestCases } from './lib/utils.js';
 
 /**
  * Test the rules.
@@ -183,89 +177,15 @@ describe('Basics', () => {
     });
 });
 
-// start an server to host doc, response to sr.url requests
-const app = express();
-app.use('/docs', express.static(pth.join(__dirname, 'docs')));
-
-// use express-handlebars
-app.engine(
-    'handlebars',
-    exphbs.engine({
-        defaultLayout: pth.join(__dirname, './doc-views/layout/spec'),
-        layoutsDir: pth.join(__dirname, './doc-views'),
-        partialsDir: pth.join(__dirname, './doc-views/partials/'),
-    })
-);
-app.set('view engine', 'handlebars');
-app.set('views', pth.join(__dirname, './doc-views'));
-
-function renderByConfig(req, res) {
-    const { rule, type } = req.query;
-    const suffix = req.params.track
-        ? `${req.params.track}/${req.params.profile}`
-        : req.params.profile;
-
-    // get data for template from json (.js)
-    const path = pth.join(
-        __dirname,
-        `./doc-views/${req.params.docType}/${suffix}.js`
-    );
-
-    // eslint-disable-next-line node/no-unsupported-features/es-syntax
-    import(path).then(module => {
-        const data = module.default;
-
-        let finalData;
-        if (!type)
-            res.send(
-                '<h1>Error: please add the parameter "type" in the URL </h1>'
-            );
-        else if (type.startsWith('good')) {
-            finalData = data[type];
-        } else {
-            if (!rule)
-                res.send(
-                    '<h1>Error: please add the parameter "rule" in the URL </h1>'
-                );
-
-            // for data causes error, make rule and the type of error specific.
-            finalData = data[rule][type];
-        }
-        res.render(pth.join(__dirname, './doc-views/layout/spec'), finalData);
-    });
-}
-
-app.get('/doc-views/:docType/:track/:profile', renderByConfig);
-app.get('/doc-views/:docType/:profile', renderByConfig);
-
-// config single redirection
-app.get('/docs/links/image/logo', (req, res) => {
-    res.redirect('/docs/links/image/logo.png');
-});
-// config single redirection to no where (404)
-app.get('/docs/links/image/logo-fail', (req, res) => {
-    res.redirect('/docs/links/image/logo-fail.png');
-});
-// config multiple redirection
-app.get('/docs/links/image/logo-redirection-1', (req, res) => {
-    res.redirect(301, '/docs/links/image/logo-redirection-2');
-});
-app.get('/docs/links/image/logo-redirection-2', (req, res) => {
-    res.redirect(307, '/docs/links/image/logo-redirection-3');
-});
-app.get('/docs/links/image/logo-redirection-3', (req, res) => {
-    res.redirect('/docs/links/image/logo.png');
-});
-
-let server;
+let testserver;
 
 before(done => {
-    server = app.listen(PORT, done);
+    testserver = app.listen(PORT, done);
 });
 
 after(done => {
-    if (server) {
-        server.close(done);
+    if (testserver) {
+        testserver.close(done);
     }
 });
 
@@ -543,60 +463,6 @@ function checkRule(tests, options) {
         });
     });
 }
-
-// ignore .DS_Store from Mac
-function listFilesOf(dir) {
-    const files = readdirSync(dir);
-    const blocklist = ['.DS_Store', 'Base.js'];
-
-    return files.filter(v => !blocklist.find(b => v.includes(b)));
-}
-
-const flat = objs => objs.reduce((acc, cur) => ({ ...acc, ...cur }), {});
-
-const buildProfileTestCases = async path => {
-    // eslint-disable-next-line node/no-unsupported-features/es-syntax
-    const { rules } = await import(path);
-    return rules;
-};
-
-const buildTrackTestCases = async path => {
-    if (lstatSync(path).isFile()) {
-        const profile = await buildProfileTestCases(path);
-        return profile;
-    }
-
-    const profiles = await Promise.all(
-        listFilesOf(path).map(async profile => ({
-            [profile]: await buildProfileTestCases(`${path}/${profile}`),
-        }))
-    );
-
-    return flat(profiles);
-};
-
-const buildDocTypeTestCases = async path => {
-    const tracks = await Promise.all(
-        listFilesOf(path).map(async track => ({
-            [track]: await buildTrackTestCases(`${path}/${track}`),
-        }))
-    );
-
-    return flat(tracks);
-};
-
-const buildBadTestCases = async () => {
-    const base = `${process.cwd()}/test/data`;
-    const docTypes = await Promise.all(
-        listFilesOf(base)
-            .filter(v => lstatSync(`${base}/${v}`).isDirectory())
-            .map(async docType => ({
-                [docType]: await buildDocTypeTestCases(`${base}/${docType}`),
-            }))
-    );
-
-    return flat(docTypes);
-};
 
 function runTestsForProfile({ docType, track, profile, rules }) {
     // Profile: CR/NOTE/RY ...
